Extracted checkpoint related code into dedicated file
