+++ src/packages/emmett/src/processors/index.ts
@@ -1,2 +1,3 @@
+export * from './checkpoints';
 export * from './inMemoryProcessors';
 export * from './processors';
